refactor: remove legacy ai analysis service
